+++ src/routes/query_parameters/query_parameters_line.rs
@@ -1,41 +1,38 @@
 use super::RequestedCwy;
 use super::OutputFormat;
 
-use serde;
+use serde::Deserialize;
 
-#[derive(serde::Deserialize, Debug)]
+#[derive(Deserialize, Debug)]
 pub struct QueryParametersLine {
+    /// Road number
 	pub road: String,
 
-	#[serde(default = "default_cwy")]
-	pub cwy: RequestedCwy,
-
+    
+    /// the starting SLK offset to slice the road network
 	pub slk_from: f32,
+    
+    /// the ending SLK offset to slice the road network
 	pub slk_to: f32,
 
-	#[serde(default = "default_offset")]
+	#[serde(default)] // use Default trait: LRS
+    /// The carriageway filter; all carriageways are included in the result by
+    /// default
+	pub cwy: RequestedCwy,
+
+	#[serde(default)] // default 0
+    /// The number of metres to offset the point or linestring from the road
+    /// centreline. If facing the direction of increasing SLK, negative values
+    /// will offset to the left, and positive values to the right.
 	pub offset:f32,
 
-	#[serde(default = "default_output_format")]
+	#[serde(default)] // use Default trait: GEOJSON
+    /// The output data format to be returned by the server
 	pub f: OutputFormat,
 
-	#[serde(default = "default_m")]
+	#[serde(default)] // default false
+    /// request that the linear referencing M coordinate should be included if
+    /// possible
 	pub m:bool,
 
-}
-
-fn default_offset() -> f32 {
-	0.0f32
-}
-
-fn default_cwy() -> RequestedCwy {
-	RequestedCwy::LRS
-}
-
-fn default_output_format() -> OutputFormat {
-	OutputFormat::GEOJSON
-}
-
-fn default_m()->bool{
-	false
 }